move card types outside
change styles prop optional
